requirements.txt
ebooklib
beautifulsoup4
Pillow
Flask
pygame